refactor: rename design dir to builder-and-design
