+++ frontend/src/features/admin-form/create/builder-and-design/components/BuilderAndDesignDrawer/CreatePageDrawerCloseButton.tsx
