+++ frontend/src/features/admin-form/create/builder-and-design/components/BuilderAndDesignPlaceholder.tsx
