+++ frontend/src/features/admin-form/create/common/components/CreatePageContent.tsx
@@ -1,8 +1,8 @@
+import BuilderAndDesign from '~features/admin-form/create/builder-and-design'
 import {
   DrawerTabs,
   useCreatePageDrawer,
 } from '~features/admin-form/create/CreatePageDrawerContext'
-import BuilderAndDesign from '~features/admin-form/create/design'
 import CreatePageLogicTab from '~features/admin-form/create/logic'
 
 export const CreatePageContent = (): JSX.Element => {
